Use tabwriter for rendering right side progressbar text
The previous plain tab approach was very brittle and resulted in
jerkiness more often than not. Tabwriter doesn't avoid the jumping
since this can't be reliably avoided (the data will always be at risk of
overflowing the preset boundaries), but we need a responsive framework
that can handle the jumps gracefully. A few minor issues aside,
tabwriter can handle this better than us writing it manually.