Replace domainr/dnsr with semihalev/sdns
dnsr has some questionable[1] design[2] decisions in that it queries
root nameservers by default and doesn't allow changing this at runtime.
This might be acceptable for a long running DNS server, but it's a
deal-breaker for a tool like k6 that often has short running times, and
should query the locally configured nameservers. We'd need some external
caching or workarounds for this dnsr limitation, which sounds like
swimming upstream.

sdns is not meant to be used as a library, and is also a standalone DNS
server. We only need the resolver "middleware" from it, which works fine
when used as a library. So far it has a better API than dnsr, and
exposes the miekg/dns objects directly, which makes various scenarios
much easier. Plus it has other interesting goodies we might want to use[3].

[1]: https://github.com/domainr/dnsr/blob/ab97e41ec5b4d59c1c6d77babe50828ba5162dd3/root.go
[2]: https://github.com/domainr/dnsr/blob/ab97e41ec5b4d59c1c6d77babe50828ba5162dd3/generate.sh
[3]: https://github.com/semihalev/sdns/tree/59b407f31f2466eb0ff241811a4a194249f98bee/middleware