Add test HTTP and DNS servers, refactor TestDNSCache
I think this is a more comprehensive test setup that can be used by both
unit tests and functional/E2E tests, and doesn't require hacky Resolver
overrides on the js.VU. The global nameservers and SetNS() is still a
problem, but I plan to clean that up once local resolution is in place.

I didn't use httpmultibin since I needed something more lightweight that
would just receive an http.HandlerFunc and listen on the specified
address. NewHTTPMultiBin uses httptest.Server, which listens on
127.0.0.1 by default, so it was easier to use a separate smaller HTTP
server that allows configuring the address.